fixed floating point exception

+++ counts_probability.cc
@@ -54,7 +54,7 @@ int main(int argc, const char *argv[]) {
     str >> has_mutation_total;
     str >> has_no_mutation_total;
     total_trios = has_mutation_total + has_no_mutation_total;
-    probability = has_mutation_total / total_trios;
+    probability = (double) has_mutation_total / total_trios;
     probabilities.push_back(probability);
   }
   f.close();
